+++ apis/server.d.ts
@@ -1,8 +1,7 @@
-/* eslint-disable @typescript-eslint/ban-types */
 import { Service, ServiceImpl } from './services'
 import { CSN } from './csn'
 import * as http from 'http'
-import * as cds from './cds'
+import type * as cds from './cds'
 import { Application, RequestHandler } from 'express'
 import { XOR } from './internal/util'
 
@@ -13,12 +12,27 @@ export const connect: {
 
   /**
 		 * Connects to a specific datasource.
+		 * @example cds.connect.to ('service')
 		 * @see [capire](https://cap.cloud.sap/docs/node.js/cds-connect#cds-connect-to)
 		 */
   to(datasource: string, options?: cds_connect_options): Promise<Service>,
 
   /**
+	 * Shortcut for 'db' as the primary database returning `cds.DatabaseService`
+	 * @example cds.connect.to ('db')
+	*/
+  to(datasource: 'db', options?: cds_connect_options): Promise<cds.DatabaseService>,
+
+  /**
+	 * Connects to a specific datasource via a Service subclass
+	 * @example cds.connect.to (ServiceClass)
+	 * @see [capire](https://cap.cloud.sap/docs/node.js/cds-connect#cds-connect-to)
+	 */
+  to<S extends Service>(datasource: {new(): S}, options?: cds_connect_options): Promise<S>,
+
+  /**
 		 * Connects to a specific datasource via options.
+		 * @example cds.connect.to ({ kind:..., impl:... })
 		 * @see [capire](https://cap.cloud.sap/docs/node.js/cds-connect#cds-connect-to)
 		 */
   to(options: cds_connect_options): Promise<Service>,
@@ -34,6 +48,7 @@ export const connect: {
 /**
 	 * The default bootstrap function as loaded from server.js
 	 */
+// eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
 export const server: Function
 
 /**