+++ apis/test.d.ts
@@ -1,8 +1,8 @@
 import { AxiosInstance } from 'axios'
-import chai from 'chai'
+import type chai from 'chai'
 import * as http from 'http'
 import { Service } from './services'
-import * as cds from './cds'
+import type * as cds from './cds'
 
 type _cds = typeof cds